Update code to better utilize plugins instead of everything native, convert dnsdb code into a plugin, plugin code is still WIP
