+++ pydat5-dev/frontend/src/components/helpers/dropdown_cell.jsx
@@ -1,0 +1,42 @@
+import React, {useState} from 'react'
+
+import ArrowDropDownIcon from '@material-ui/icons/ArrowDropDown';
+import IconButton from '@material-ui/core/IconButton'
+import Menu from '@material-ui/core/Menu';
+
+
+export const DropDownCell = (props) => {
+    const [anchorEl, setAnchorEl] = useState(null)
+
+    const handleClick = (e) => {
+        setAnchorEl(e.currentTarget)
+    }
+
+    const handleClose = () => {
+        setAnchorEl(null)
+    }
+
+    return (
+        <React.Fragment>
+            <IconButton
+                aria-controls={`${props.friendly}-menu`}
+                onClick={handleClick}
+                size='small'
+            >
+                <ArrowDropDownIcon />
+            </IconButton>
+            <Menu
+                anchorEl={anchorEl}
+                keepMounted
+                open={Boolean(anchorEl)}
+                onClose={handleClose}
+            >
+                {props.children}
+            </Menu>
+            {props.value}
+        </React.Fragment>
+
+    )
+}
+
+export default DropDownCell
