+++ pydat5-dev/frontend/src/components/helpers/preferences.jsx
@@ -44,7 +44,7 @@ class UserPreferencesContainer {
             return inprefs
         }
 
-        console.log(outprefs)
+        // console.log(outprefs)
         return outprefs
     }
