+++ pydat5-dev/frontend/src/components/layout/dashboard_items.jsx
@@ -3,19 +3,16 @@ import {useHistory} from 'react-router-dom'
 import ListItem from '@material-ui/core/ListItem';
 import ListItemIcon from '@material-ui/core/ListItemIcon';
 import ListItemText from '@material-ui/core/ListItemText';
-import ListSubheader from '@material-ui/core/ListSubheader';
 import Divider from '@material-ui/core/Divider'
 
 import SearchIcon from '@material-ui/icons/Search'
-import LanguageIcon from '@material-ui/icons/Language'
 import HelpIcon from '@material-ui/icons/Help'
 import EqualizerIcon from '@material-ui/icons/Equalizer'
 
-import {PyDatPluginContext} from '../plugins'
-import { useContext } from 'react';
+import {PluginManagers} from '../plugins'
 
 export const MainListItems = (props) => {
-  const drawer_plugins = useContext(PyDatPluginContext).drawer
+  const drawer_plugins = PluginManagers.drawer
 
   let history = useHistory()
 
@@ -29,10 +26,6 @@ export const MainListItems = (props) => {
       <ListItem button onClick={() => {handleRedirect('/whois')}}>
         <ListItemIcon> <SearchIcon /> </ListItemIcon>
         <ListItemText primary="WHOIS Search" />
-      </ListItem>
-      <ListItem button onClick={() => {handleRedirect('/passive')}}>
-        <ListItemIcon> <LanguageIcon /> </ListItemIcon>
-        <ListItemText primary="Passive DNS" />
       </ListItem>
       <ListItem button>
         <ListItemIcon> <EqualizerIcon /> </ListItemIcon>
