+++ pydat5-dev/frontend/src/pydat.jsx
@@ -1,4 +1,4 @@
-import React, {Suspense, useContext} from 'react'
+import React, {Suspense} from 'react'
 import {
     BrowserRouter as Router,
     Switch,
@@ -9,21 +9,21 @@ import {
 import { ThemeProvider } from '@material-ui/styles'
 import useMediaQuery from '@material-ui/core/useMediaQuery'
 
-import {PluginManagers, PyDatPluginContext} from './components/plugins'
+import './plugins'
+
+import {PluginManagers} from './components/plugins'
 import {UserPreferences, UserPreferencesContext} from './components/helpers/preferences'
 import Dashboard from './components/layout/dashboard'
 import NotFound from './components/layout/notfound'
 import {BackdropLoader} from './components/helpers/loaders'
 import { defaultTheme, darkTheme } from './components/layout/themes'
 
-import './plugins'
 
 
 const WhoisHandler = React.lazy(() => import('./components/whois'))
-const PassiveHandler = React.lazy(() => import ('./components/passive'))
 
 const Pydat = () => {
-    const routes = useContext(PyDatPluginContext).routes
+    const routes = PluginManagers.routes
     const enableDarkMode = useMediaQuery('@media (prefers-color-scheme: dark')
 
     const theme = enableDarkMode ? darkTheme : defaultTheme
@@ -32,44 +32,39 @@ const Pydat = () => {
         <React.Fragment>
             <UserPreferencesContext.Provider value={UserPreferences}>
                 <ThemeProvider theme={theme}>
-                    <PyDatPluginContext.Provider value={PluginManagers}>
-                        <Router>
-                            <Dashboard>
-                                <Suspense fallback={<BackdropLoader/>}>
-                                    <Switch>
-                                        <Route exact path="/">
-                                            <Redirect to="/whois" />
-                                        </Route>
-                                        <Route path="/whois">
-                                            <WhoisHandler/>
-                                        </Route>
-                                        <Route path="/passive">
-                                            <PassiveHandler />
-                                        </Route>
-                                        <Route path="/help">
+                    <Router>
+                        <Dashboard>
+                            <Suspense fallback={<BackdropLoader/>}>
+                                <Switch>
+                                    <Route exact path="/">
+                                        <Redirect to="/whois" />
+                                    </Route>
+                                    <Route path="/whois">
+                                        <WhoisHandler/>
+                                    </Route>
+                                    <Route path="/help">
 
-                                        </Route>
-                                        <Route path="/stats">
+                                    </Route>
+                                    <Route path="/stats">
 
-                                        </Route>
-                                        {Object.keys(routes.plugins).map((name, index) => {
-                                            return (
-                                                <Route
-                                                    key={index}
-                                                    exact
-                                                    path={routes.plugins[name].path}
-                                                    {...routes.plugins[name].extra}
-                                                >
-                                                    {React.cloneElement(routes.plugins[name].component)}
-                                                </Route>
-                                            )
-                                        })}
-                                        <Route component={NotFound} />
-                                    </Switch>
-                                </Suspense>
-                            </Dashboard>
-                        </Router>
-                    </PyDatPluginContext.Provider>
+                                    </Route>
+                                    {Object.keys(routes.plugins).map((name, index) => {
+                                        return (
+                                            <Route
+                                                key={index}
+                                                exact
+                                                path={routes.plugins[name].path}
+                                                {...routes.plugins[name].extra}
+                                            >
+                                                {React.cloneElement(routes.plugins[name].component)}
+                                            </Route>
+                                        )
+                                    })}
+                                    <Route component={NotFound} />
+                                </Switch>
+                            </Suspense>
+                        </Dashboard>
+                    </Router>
                 </ThemeProvider>
             </UserPreferencesContext.Provider>
         </React.Fragment>
